requirements.txt
aiogram
python-docx
PyPDF2
pytesseract
pillow
pandas
openpyxl
